# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the SITCON Camp 2026 game.",
        "title": "Camp 2026 Game API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/matches/join": {
            "post": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Joins a waiting two-player quiz match room by invite code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Join match room",
                "parameters": [
                    {
                        "description": "Join match request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/matches.JoinMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/matches.JoinMatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/matches/{matchID}/answers": {
            "post": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Accepts the authenticated player's answer for the current question. Correctness is revealed to both players when the round enters the reveal phase.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Submit match answer",
                "parameters": [
                    {
                        "description": "Answer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/matches.AnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/matches.AnswerAcceptedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/matches/{matchID}/loadout": {
            "put": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Updates the authenticated player's sitone loadout for a waiting match and saves it as the player's default loadout.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Update match sitone loadout",
                "parameters": [
                    {
                        "description": "Update loadout request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/matches.UpdateLoadoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/matches.UpdateLoadoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorDetail": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ProblemDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpx.ErrorDetail"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "matches.AnswerAcceptedResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "matches.AnswerRequest": {
            "type": "object",
            "required": [
                "choice",
                "questionId"
            ],
            "properties": {
                "choice": {
                    "type": "string",
                    "enum": [
                        "A",
                        "B",
                        "C",
                        "D"
                    ],
                    "example": "A"
                },
                "questionId": {
                    "type": "string",
                    "example": "quiz-001"
                }
            }
        },
...
        "matches.JoinMatchRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 4,
                    "example": "ABC123"
                }
            }
        },
        "matches.JoinMatchResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ABC123"
                },
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentQuestion": {
                    "$ref": "#/definitions/matches.MatchQuestionResponse"
                },
                "currentQuestionIndex": {
                    "type": "integer",
                    "example": 0
                },
                "currentQuestionResult": {
                    "$ref": "#/definitions/matches.MatchQuestionResult"
                },
                "hostPlayerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "matchId": {
                    "type": "string",
                    "example": "match_7H9K2Q"
                },
                "mode": {
                    "type": "string",
                    "example": "pvp"
                },
                "phase": {
                    "type": "string",
                    "example": "answering"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/matches.MatchPlayerResponse"
                    }
                },
                "questionCount": {
                    "type": "integer",
                    "example": 10
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/matches.MatchQuestionResult"
                    }
                },
                "revealEndsAt": {
                    "type": "string"
                },
                "roundEndsAt": {
                    "type": "string"
                },
                "roundStartedAt": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "matches.MatchItemDropResponse": {
            "type": "object",
            "properties": {
                "dropRate": {
                    "type": "integer",
                    "example": 60
                },
                "dropped": {
                    "type": "boolean",
                    "example": true
                },
                "itemId": {
                    "type": "string",
                    "example": "item_clean_spec"
                },
                "itemName": {
                    "type": "string",
                    "example": "整潔規格書"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "matches.MatchPlayerResponse": {
            "type": "object",
            "properties": {
                "answerScoreBonusPercent": {
                    "type": "integer",
                    "example": 10
                },
                "answeredCurrentQuestion": {
                    "type": "boolean",
                    "example": true
                },
                "baseOpenPowerReward": {
                    "type": "integer",
                    "example": 105
                },
                "eliminateChancePercent": {
                    "type": "integer",
                    "example": 25
                },
                "eliminateCount": {
                    "type": "integer",
                    "example": 1
                },
                "eliminatedBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "靈光型小石"
                    ]
                },
                "eliminatedChoices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "B",
                        "C"
                    ]
                },
                "kind": {
                    "type": "string",
                    "example": "human"
                },
                "materialDrop": {
                    "$ref": "#/definitions/matches.MatchItemDropResponse"
                },
                "materialDropBonusPercent": {
                    "type": "integer",
                    "example": 15
                },
                "maxScore": {
                    "type": "integer",
                    "example": 2250
                },
                "nickname": {
                    "type": "string",
                    "example": "Alice"
                },
                "openPowerBonusPercent": {
                    "type": "integer",
                    "example": 20
                },
                "openPowerReward": {
                    "type": "integer",
                    "example": 126
                },
                "playerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "ready": {
                    "type": "boolean",
                    "example": true
                },
                "score": {
                    "type": "integer",
                    "example": 850
                },
                "sitoneIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "stone_engineering_base",
                        "stone_explorer_base"
                    ]
                }
            }
        },
        "matches.MatchQuestionResponse": {
            "type": "object",
            "properties": {
                "choiceA": {
                    "type": "string",
                    "example": "git init"
                },
                "choiceB": {
                    "type": "string",
                    "example": "git clone"
                },
                "choiceC": {
                    "type": "string",
                    "example": "git status"
                },
                "choiceD": {
                    "type": "string",
                    "example": "git add"
                },
                "prompt": {
                    "type": "string",
                    "example": "哪個指令可以初始化新的 Git 儲存庫？"
                },
                "questionId": {
                    "type": "string",
                    "example": "quiz-001"
                }
            }
        },
        "matches.MatchQuestionResult": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/matches.MatchAnswerResponse"
                    }
                },
                "choiceA": {
                    "type": "string",
                    "example": "git init"
                },
                "choiceB": {
                    "type": "string",
                    "example": "git clone"
                },
                "choiceC": {
                    "type": "string",
                    "example": "git status"
                },
                "choiceD": {
                    "type": "string",
                    "example": "git add"
                },
                "correctChoice": {
                    "type": "string",
                    "example": "A"
                },
                "explanation": {
                    "type": "string",
                    "example": "git init 會在目前目錄建立新的 Git 儲存庫。"
                },
                "prompt": {
                    "type": "string",
                    "example": "哪個指令可以初始化新的 Git 儲存庫？"
                },
                "questionId": {
                    "type": "string",
                    "example": "quiz-001"
                }
            }
        },
...
        "matches.UpdateLoadoutRequest": {
            "type": "object",
            "required": [
                "sitoneIds"
            ],
            "properties": {
                "sitoneIds": {
                    "type": "array",
                    "maxItems": 5,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "stone_engineering_base",
                        "stone_explorer_base"
                    ]
                }
            }
        },
        "matches.UpdateLoadoutResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ABC123"
                },
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentQuestion": {
                    "$ref": "#/definitions/matches.MatchQuestionResponse"
                },
                "currentQuestionIndex": {
                    "type": "integer",
                    "example": 0
                },
                "currentQuestionResult": {
                    "$ref": "#/definitions/matches.MatchQuestionResult"
                },
                "hostPlayerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "matchId": {
                    "type": "string",
                    "example": "match_7H9K2Q"
                },
                "mode": {
                    "type": "string",
                    "example": "pvp"
                },
                "phase": {
                    "type": "string",
                    "example": "answering"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/matches.MatchPlayerResponse"
                    }
                },
                "questionCount": {
                    "type": "integer",
                    "example": 10
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/matches.MatchQuestionResult"
                    }
                },
                "revealEndsAt": {
                    "type": "string"
                },
                "roundEndsAt": {
                    "type": "string"
                },
                "roundStartedAt": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthCookieAuth": {
            "description": "User auth cookie. Send as `Cookie: camp2026_auth=\u003cbase64url-auth-token\u003e`.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}